+++ lib/modl/parser/interpreter.rb
@@ -11,20 +11,14 @@ module MODL
   class InterpreterError < StandardError
   end
 
+  class ParserError < StandardError
+  end
+
   # This is the main Ruby Interpreter entry point. Supply a String containing MODL text and it will return a String
   # containing the JSON equivalent. The JSON isn't pretty-printed unless pretty is true
   class Interpreter
     def self.interpret(str, pretty = false)
-      # Parse the MODL string into a MODL::Parser::Parsed object.
-      parsed = MODL::Parser::Parser.parse str
-
-      # Convert the Parsed object into a simpler structure of and Array or Hash
-      interpreted = parsed.extract_hash
-
-      # Process any class definitions used by the MODL file.
-      MODL::Parser::ClassProcessor.process(parsed.global, interpreted)
-      MODL::Parser::InstructionProcessor.process(parsed.global, interpreted)
-      # If the result is a simple string then just return it.
+      interpreted = MODL.parse(str)
       return interpreted if interpreted.is_a? String
 
       # Otherwise generate a JSON string.